Feat : 타이틀 게임방법 룰 추가, 폭탄 쿨타임 추가

+++ Kung/Assets/Scripts/GameUI/Shortcut/ShortcutKey.cs
@@ -11,6 +11,8 @@ public class ShortcutKey : MonoBehaviour
     [SerializeField] private GameObject _dynamite;
     [SerializeField] private Tilemap _rockTile;
 
+    private float bombCooldownTime = 3f;
+    private float cooldownTime;
 
 
     private ShortCutSlotUI[] _slots;
@@ -33,52 +35,62 @@ public class ShortcutKey : MonoBehaviour
             _slots[i].SetData(viewData);
         }
     }
-    // 폭탄이랑 빅폭탄은 누르면 현재위치에 설치 (isGround)체크해야함
-    // 다이나마이트는 누르면 현재위치 바로 아래에 설치 (isGround)체크해야함
-    // 아래가 돌인지 감지하고 그 돌위치에 소환
 
-    // 구급상자는 플레이어 Health.hp 이런데 접근해서 100 힐
-    // 산소는 플레이어 Air.air 이런데 접근해서 100% 힐
+    private void Update()
+    {
+        cooldownTime += Time.deltaTime;
+    }
 
-    // 모든 버튼들은 누른 뒤에 지금 해당 아이템을 가지고있는지. 수량 확인 해야함
     public void OnBombButtonClick()
     {
-        if (_shortCutServiceLocator.Service.CanUseShortCutItem(1003) && _groundChecker.IsGrounded)
+        if(cooldownTime > bombCooldownTime)
         {
-            _shortCutServiceLocator.Service.RemoveShortCutItemQuantity(1003);
-            Instantiate(_bomb, _playerTransform.position, Quaternion.identity);
-            Refresh();
+            if (_shortCutServiceLocator.Service.CanUseShortCutItem(1003) && _groundChecker.IsGrounded)
+            {
+                _shortCutServiceLocator.Service.RemoveShortCutItemQuantity(1003);
+                Instantiate(_bomb, _playerTransform.position, Quaternion.identity);
+                Refresh();
+            }
+            cooldownTime = 0f;
         }
     }
     public void OnBigBombButtonClick()
     {
-        if (_shortCutServiceLocator.Service.CanUseShortCutItem(1004) && _groundChecker.IsGrounded)
+        if (cooldownTime > bombCooldownTime)
         {
-            _shortCutServiceLocator.Service.RemoveShortCutItemQuantity(1004);
-            Instantiate(_bigBomb, _playerTransform.position, Quaternion.identity);
-            Refresh();
+            if (_shortCutServiceLocator.Service.CanUseShortCutItem(1004) && _groundChecker.IsGrounded)
+            {
+                _shortCutServiceLocator.Service.RemoveShortCutItemQuantity(1004);
+                Instantiate(_bigBomb, _playerTransform.position, Quaternion.identity);
+                Refresh();
+            }
+            cooldownTime = 0f;
         }
     }
     public void OnDynamiteButtonClick()
     {
-        if (_groundChecker.IsGrounded == true)
+        if (cooldownTime > bombCooldownTime)
         {
-            // 현재 플레이어의 위치 밑의 미네랄 타일을 가져와야함
-            Vector3Int DownCell = _rockTile.WorldToCell(_playerTransform.position + new Vector3(0, -0.1f, 0));
-            TileBase tile = _rockTile.GetTile(DownCell);
-
-            if (tile is CustomOreTile customOreTile)
+            if (_groundChecker.IsGrounded == true)
             {
-                if (customOreTile.oreType == CustomOreTile.OreType.Rock)
+                // 현재 플레이어의 위치 밑의 미네랄 타일을 가져와야함
+                Vector3Int DownCell = _rockTile.WorldToCell(_playerTransform.position + new Vector3(0, -0.1f, 0));
+                TileBase tile = _rockTile.GetTile(DownCell);
+
+                if (tile is CustomOreTile customOreTile)
                 {
-                    if (_shortCutServiceLocator.Service.CanUseShortCutItem(1005))
+                    if (customOreTile.oreType == CustomOreTile.OreType.Rock)
                     {
-                        _shortCutServiceLocator.Service.RemoveShortCutItemQuantity(1005);
-                        Instantiate(_dynamite, _rockTile.CellToWorld(DownCell) + new Vector3(0.15f, 0.15f, 0), Quaternion.identity);
-                        Refresh();
+                        if (_shortCutServiceLocator.Service.CanUseShortCutItem(1005))
+                        {
+                            _shortCutServiceLocator.Service.RemoveShortCutItemQuantity(1005);
+                            Instantiate(_dynamite, _rockTile.CellToWorld(DownCell) + new Vector3(0.15f, 0.15f, 0), Quaternion.identity);
+                            Refresh();
+                        }
                     }
                 }
             }
+            cooldownTime = 0f;
         }
     }
     public void OnFirstAidKitButtonClick()
